use adb shell command from API25+ for FontScaleRule
